+++ Extensions/Header.h
@@ -17,8 +17,8 @@ namespace ADTFStream::Extensions
 		unsigned int userId = 0;
 		unsigned int typeId = 0;
 		unsigned int versionId = 0;
-		unsigned long dataOffset = 0;
-		unsigned long dataSize = 0;
+		unsigned long long dataOffset = 0;
+		unsigned long long dataSize = 0;
 		unsigned char reserved[96];
 
 		// Constructors
