+++ Extensions/MarkerInfo.cpp
@@ -3,7 +3,7 @@
 
 using ADTFStream::Extensions::MarkerInfo;
 
-MarkerInfo::MarkerInfo(FILE* file, unsigned long dataSize)
+MarkerInfo::MarkerInfo(FILE* file, unsigned long long dataSize)
 {
 	markers = new char[dataSize];
 	fread(markers, dataSize, 1, file);
