+++ Extensions/ReferencedFiles.cpp
@@ -5,7 +5,7 @@
 
 using ADTFStream::Extensions::ReferencedFiles;
 
-ReferencedFiles::ReferencedFiles(FILE* file, unsigned long dataSize)
+ReferencedFiles::ReferencedFiles(FILE* file, unsigned long long dataSize)
 {
 	data = new char[dataSize];
 	fread(data, dataSize, 1, file);
