+++ Extensions/StreamIndex.h
@@ -14,9 +14,9 @@ namespace ADTFStream::Extensions
 		{
 		  public:
 			// Member variables
-			unsigned long fileIndexEntryCount = 0;
-			unsigned long firstTime = 0;
-			unsigned long lastTime = 0;
+			unsigned long long fileIndexEntryCount = 0;
+			unsigned long long firstTime = 0;
+			unsigned long long lastTime = 0;
 			unsigned int dataSize = 0;
 			char name[228];
 		};
@@ -40,11 +40,11 @@ namespace ADTFStream::Extensions
 		unsigned int codecSize = 0;
 		unsigned char* codec = nullptr;
 
-		unsigned long entryCount = 0;
+		unsigned long long entryCount = 0;
 		unsigned int* entries = nullptr;
 
 		// Constructor
-		StreamIndex(FILE* file, unsigned long dataSize);
+		StreamIndex(FILE* file, unsigned long long dataSize);
 
 		// Destructor
 		~StreamIndex();
